Editorial: document potential error in selectRange
Prior to this commit, the algorithm for
`Intl.PluralRules.prototype.selectRange` asserted that its invocation of
the ResolvePluralRange abstract operation would not return an abrupt
completion. This was inaccurate because ResolvePluralRange returns a
throw completion when the two input numbers do not describe a valid
range and because because the relationship between the specification
variables `x` and `y` is not verified prior to invocation.

Communicate the potential for failure by replacing the `!` shorthand
with `?`.